Move BoolSetting next to other Settings utilities since it is not a Value
